requirements.txt
asgiref==3.2.10
boto3==1.15.2
botocore==1.18.2
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
django-environ==0.4.5
django-filter==2.3.0
django-on-heroku==1.0.1
django-storages==1.10.1
djangorestframework==3.11.1
jmespath==0.10.0
Markdown==3.2.2
Pillow==7.2.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
gunicorn==20.0.4
